feat: adds way of parsing dat files, insert values in features and features_data
