+++ api/apps/geoprocessing/src/legacy-project-import/legacy-piece-importers/legacy-piece-importers.module.ts
@@ -13,7 +13,7 @@ import { Logger, Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { AppConfig } from '../../utils/config.utils';
 import { FileReadersModule } from './file-readers/file-readers.module';
-import { FeaturesLegacyProjectPieceImporter } from './features.legacy-project-piece-importer';
+import { FeaturesLegacyProjectPieceImporter } from './features.legacy-piece-importer';
 import { PlanningGridLegacyProjectPieceImporter } from './planning-grid.legacy-piece-importer';
 import { ScenarioPusDataLegacyProjectPieceImporter } from './scenarios-pus-data.legacy-piece-importer';
